Define Hyprland bindings for taking screenshots
I'd like to use gnome-screenshot, as when capturing regions it remembers
which region you just captured. This is *very* useful when capturing
screenshots of slides during presentations.

But it crashes when I run it. So hyprshot will do for now.

The --silent switch prevents a notification message from popping up to
confirm the screenshot was taken. The --freeze switch freezes the screen
so you can definitely capture what you were after.