Remove inactive window border
This change was inspired by how it works on Pop!_OS 22.04, which I
really like. Only the active window gets a border, which makes it that
little bit easier to see which window is active when borders are thin
and gaps are small.

This solution came from this thread (see the comments for info if you
want to change how this is implemented):

https://github.com/hyprwm/Hyprland/issues/4631