Style wofi launcher using Catppuccin colours
I followed the advice in Diego Sandoval's video [1]. Colours for
Catppuccin were taken from [2].

See wofi(5) for information on the config settings and the CSS selectors
that are available.

Note that icons are disabled, but I've defined an image size of 24px and
have added some styles to target them. Having worked out how to do it I
thought I ought to check that code in now, in case I later learn that I
miss icons.

[1] https://www.youtube.com/watch?v=KfOYXvRMcIQ
[2] https://github.com/quantumfate/wofi