Launch programs in Hyprland via uwsm
I've added a `~/.bashrc-arch` script (that my `.bashrc` will source on
Arch machine) that automatically launches Hyprland via uwsm, when I
login.

uwsm can wrap all the programs we start in some systemd goodness, so
that each app we start can be treated as a unit.

Any apps that we launch this way run as children of systemd, rather than
as children of the compositor.

Note that I've not launched wofi via uwsm; it'll be a child process of
Hyprland. This feels like a good fit for such a short-lived program.

For more info see the Hyprland wiki:

https://wiki.hypr.land/Useful-Utilities/Systemd-start/#launching-applications-inside-session