Set GTK theme
The theme is installed on Arch by a package in the AUR [1], which is
installed by the Hyprland playbook in my infra repository.

You can select it with nwg-look (also installed via that same playbook),
which will persist the settings in some config files that it'll create
inside ~/.config. This approach feels much easier to manage; we don't
need to version those auto-generated config files for consistency across
different machines.

[1] catppuccin-gtk-theme-mocha