Rename Hyprland special workspace to "scratch"
You can have multiple special workspaces, and need to give each one a
name. "magic" is just what my first one was called in the example
config.

I'm aiming to use it as a general purpose workspace for windows that
it's useful to have on hand most of the time, and Super+s is easy to
access. So I'll call it "scratch" to help me remember that bind.