Launch GNOME Pomodoro on scratch Hyprland workspace
I toggle this on/off with the Super+s keybinding. With luck, I won't put
much other stuff on that workspace. But it's in case I do that I've used
pseudo rather than float here, as it'll allow new windows to push GNOME
Pomodoro to one side, rather than covering it up/being covered by it.