Run Obsidian on "notes" workspace in Hyprland
I've been finding it a bit odd pressing Super+s to get to Obsidian. It's
not that easy to do with one hand. I considered using Super+o but I've
got that mapped to toggle horizontal/vertical splits, and it's in my
muscle memory from Pop.

I'm liking Super+n and "notes" as it'll make sense when using other note
taking apps too.