Set RIPGREP_CONFIG_PATH via direnv
I'd initially set RIPGREP_CONFIG_PATH globally, in one of my .bashrc
files.

It turns out that if the config file specified in RIPGREP_CONFIG_PATH
doesn't exist, ripgrep will complain about it. So I should only set the
variable to a relative path where the file exists, and direnv can do
that for us.

Run `direnv allow` when cd-ing into the directory for the first time,
then direnv will set/unset the directory as we cd into it/out of it.